+++ Sources/KeyboardKit/_Deprecated/InputSetProvider+Deprecated.swift
@@ -3,7 +3,7 @@ public extension InputSetProvider {
     /**
      Create an input row from a string.
      */
-    @available(*, deprecated, message: "Use InputSet.init(chars:) instead")
+    @available(*, deprecated, message: "Use InputSetRow(chars:) instead")
     func row(_ chars: String) -> InputSetRow {
         InputSetRow(chars: chars.chars)
     }
@@ -11,7 +11,7 @@ public extension InputSetProvider {
     /**
      Create an input row from a char array.
      */
-    @available(*, deprecated, message: "Use InputSet.init(chars:) instead")
+    @available(*, deprecated, message: "Use InputSetRow(chars:) instead")
     func row(_ chars: [String]) -> InputSetRow {
         InputSetRow(chars: chars)
     }
@@ -22,12 +22,15 @@ public extension InputSetProvider {
 
      Both arrays must contain the same amount of characters.
      */
-    @available(*, deprecated, message: "Use InputSetRow.init(lowercased:uppercased:) instead")
+    @available(*, deprecated, message: "Use InputSetRow(lowercased:uppercased:) instead")
     func row(
         lowercased: String,
-        uppercased: String) -> InputSetRow {
-        row(lowercased: lowercased.chars,
-            uppercased: uppercased.chars)
+        uppercased: String
+    ) -> InputSetRow {
+        InputSetRow(
+            lowercased: lowercased.chars,
+            uppercased: uppercased.chars
+        )
     }
 
     /**
@@ -36,12 +39,14 @@ public extension InputSetProvider {
 
      Both arrays must contain the same amount of characters.
      */
-    @available(*, deprecated, message: "Use InputSetRow.init(lowercased:uppercased:) instead")
+    @available(*, deprecated, message: "Use InputSetRow(lowercased:uppercased:) instead")
     func row(
         lowercased: [String],
-        uppercased: [String]) -> InputSetRow {
+        uppercased: [String]
+    ) -> InputSetRow {
         InputSetRow(
             lowercased: lowercased,
-            uppercased: uppercased)
+            uppercased: uppercased
+        )
     }
 }